Fix LOT_SIZE error in calc_quantity_from_part() (Issue #26).  move get_part() and calc_quantity_from_part() to BinanceExchange.py.
