+++ ExchangeInterfaces/Exchange.py
@@ -35,32 +35,6 @@ class Exchange:
     def get_balance_coin_by_symbol(self, symbol):
         return list(filter(lambda el: el['asset'] == symbol[:3], self.get_balance()))[0]
 
-    def get_part(self, symbol, quantity, price, side):
-        # get part of the total balance of this coin
 
-        # if order[side] == sell: need obtain coin balance
-        if side == 'BUY':
-            balance = float(self.get_balance_market_by_symbol(symbol)['free'])
-            part = float(quantity)*float(price)/balance
-        else:
-            balance = float(self.get_balance_coin_by_symbol(symbol)['free'])
-            part = float(quantity)/balance
 
-        part = part * 0.99  # decrease part for 1% for avoid rounding errors in calculation
-        return part
 
-    def calc_quantity_from_part(self, symbol, quantityPart, price, side):
-        # calculate quantity from quantityPart
-
-        # if order[side] == sell: need obtain coin balance
-        if side == 'BUY':
-            cur_bal = float(self.get_balance_market_by_symbol(symbol)['free'])
-            quantity = float(quantityPart) * float(cur_bal) / float(price)
-        else:
-            cur_bal = float(self.get_balance_coin_by_symbol(symbol)['free'])
-            quantity = quantityPart*cur_bal
-
-        # balanceIndex = [idx for idx, element in enumerate(self.get_balance()) if element['asset'] == str(symbol)[3:]][0]
-        # cur_bal = float(self.get_balance()[balanceIndex]['free'])
-        quantity = round(quantity, 6)
-        return quantity
